refactor(db): extract app_state counter to db/src/app_state.rs
