+++ db/src/lib.rs
@@ -4,6 +4,7 @@
 //! (axum REST handlers) and `frontend/` (Dioxus server functions under
 //! `feature = "server"`).
 
+pub mod app_state;
 pub mod auth;
 pub mod author_photos;
 pub mod ebook;
@@ -18,6 +19,7 @@ pub mod worker;
 // Flatten the query layer so callers write `omnibus_db::list_books(...)`
 // instead of `omnibus_db::queries::list_books(...)`. Keeps callsites terse
 // and mirrors how `db.rs` looked before the extraction.
+pub use app_state::*;
 pub use pool::*;
 pub use queries::*;
 pub use thumbs::{thumb_path_for, thumbs_dir, ThumbError, ThumbSize};
